refactor: components -> editors
