+++ packages/presets/src/index.ts
@@ -1,4 +1,4 @@
-export * from './components/index.js';
+export * from './editors/index.js';
 
 const env =
   typeof globalThis !== 'undefined'
